+++ InterfaceLib/InterFaces/ITeamContainer.cs
@@ -2,11 +2,11 @@
 {
     public interface ITeamContainer
     {
-        public TeamDTO FindByID(long id);
+        public TeamDTO FindByID(int id);
 
         public List<TeamDTO> GetAllTeams();
 
-        public List<TeamDTO> GetAllTeamsFromClub(long clubID);
+        public List<TeamDTO> GetAllTeamsFromClub(int clubID);
 
     }
 }
